Spam Bug

+++ bot.js
@@ -15,28 +15,29 @@ const isValidCommand = (message, cmdName) => message.content.toLowerCase().start
 
 //Enforce Nickname On Welcome
 client.on('message', function(message) {
-
-
-    if (!message.guild.me.hasPermission('MANAGE_NICKNAMES')) return message.channel.send('I don\'t have permission to change your nickname!');
-    message.member.setNickname( message.author.username + " | " + message.content , " ").catch(err=>console.log(err));
-    let { cache } = message.guild.roles;
-    let role = cache.find(role => role.name.toLowerCase() === MemberRole);
-    if(role) {
-        if(message.member.roles.cache.has(role.id)){
+    if(message.author.bot) return;
+    if(message.channel.name === WelcomeChannel){
+        if (!message.guild.me.hasPermission('MANAGE_NICKNAMES')) return message.channel.send('I don\'t have permission to change your nickname!');
+        message.member.setNickname( message.author.username + " | " + message.content , " ").catch(err=>console.log(err));
+        let { cache } = message.guild.roles;
+            let role = cache.find(role => role.name.toLowerCase() === MemberRole);
+        if(role) {
+            if(message.member.roles.cache.has(role.id)){
             message.channel.send("You already have this role!");
-        }
+            }
+            else{
+                message.channel.send("IGN added!")
+                setTimeout(() => {
+                    message.channel.bulkDelete(100);
+                     message.member.roles.add(role);
+                }, 2000);
+                setTimeout(() => {
+                    message.channel.send("Please enter your IGN to have full access of the server!<33")
+                }, 4000);
+            }
+       }
         else{
-            message.channel.send("IGN added!")
-            setTimeout(() => {
-                message.channel.bulkDelete(100);
-                message.member.roles.add(role);
-            }, 2000);
-            setTimeout(() => {
-                message.channel.send("Please enter your IGN to have full access of the server!<33")
-            }, 4000);
+            message.channel.send("Role not found!")
         }
-    }
-    else{
-        message.channel.send("Role not found!")
     }
 });
